adding card controller et dao

+++ Atelier1/WebDynamique/src/main/java/fr/cpe/scoobygang/atelier1/model/Card.java
@@ -1,5 +1,6 @@
 package fr.cpe.scoobygang.atelier1.model;
 
+import jakarta.annotation.Nullable;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -16,9 +17,9 @@ public class Card {
     private String smallImgUrl;
     private int id;
     private int energy;
-    private float hp;
-    private float defence;
-    private float attack;
-    private float price;
-    private int userId;
+    private double hp;
+    private double defence;
+    private double attack;
+    private double price;
+    private Integer userId;
 }
